backend/pyproject.toml
[project]
name = "policyrag"
version = "0.1.0"
description = "RAG system for SEC financial filing QA with hallucination scoring"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "chromadb>=0.4.24",
    "sentence-transformers>=2.5.0",
    "openai>=1.12.0",
    "groq>=0.9.0",
    "google-generativeai>=0.7.0",
    "langchain-text-splitters>=0.0.1",
    "pymupdf>=1.23.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
